requirements.txt
langchain-community
tqdm
pypdf
langchain-openai
faiss-cpu
langchain
python-dotenv